# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface MobileBotSignature {
  ownerWallet: string;
  nonce: number;
  signature: string;
}

export type MobileSwingSummaryMode = typeof MobileSwingSummaryMode[keyof typeof MobileSwingSummaryMode];


export const MobileSwingSummaryMode = {
  paper: 'paper',
  live: 'live',
} as const;

export interface MobileSwingSummary {
  id: string;
  name: string;
  mode: MobileSwingSummaryMode;
  status: string;
  workerWallet?: string;
  /** @nullable */
  targetMint?: string | null;
  maxPositions?: number;
  minLiquidityUsd?: number;
  /** @nullable */
  minPoolAgeHours?: number | null;
  requireApproval?: boolean;
  /** @nullable */
  takeProfitPct?: number | null;
  profitSkimPct?: number;
  lossStopCount?: number;
  [key: string]: unknown;
 }

export interface MobileSwingCreated {
  strategy: MobileSwingSummary;
  fundingAddress: string;
  recommendedFundingSol: number;
}

export type MobileSwingSettingsRedEndBehavior = typeof MobileSwingSettingsRedEndBehavior[keyof typeof MobileSwingSettingsRedEndBehavior];


export const MobileSwingSettingsRedEndBehavior = {
  sell: 'sell',
  send_tokens: 'send_tokens',
} as const;

export type MobileSwingSettings = MobileBotSignature & ({
  /**
     * @minLength 1
     * @maxLength 40
     */
  name?: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  maxPositions?: number;
  /**
     * @minimum 5000
     * @maximum 1000000
     */
  minLiquidityUsd?: number;
  /**
     * @minimum 0
     * @maximum 720
     */
  minPoolAgeHours?: number;
  requireApproval?: boolean;
  /**
     * @minimum 3
     * @maximum 500
     * @nullable
     */
  takeProfitPct?: number | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  profitSkimPct?: number;
  /**
     * @minimum 1
     * @maximum 20
     */
  lossStopCount?: number;
  /**
     * @minimum 2
     * @maximum 50
     * @nullable
     */
  stopRoomPct?: number | null;
  /**
     * @minimum 40
     * @maximum 90
     * @nullable
     */
  winnerKeepPct?: number | null;
  neverSellAtLoss?: boolean;
  redEndBehavior?: MobileSwingSettingsRedEndBehavior;
  allowStacking?: boolean;
  hotStreakFullSize?: boolean;
});

export type MobileSwingCreateMode = typeof MobileSwingCreateMode[keyof typeof MobileSwingCreateMode];


export const MobileSwingCreateMode = {
  paper: 'paper',
  live: 'live',
} as const;

export type MobileSwingCreateUniverse = typeof MobileSwingCreateUniverse[keyof typeof MobileSwingCreateUniverse];


export const MobileSwingCreateUniverse = {
  crypto: 'crypto',
} as const;

export type MobileSwingCreateStyle = typeof MobileSwingCreateStyle[keyof typeof MobileSwingCreateStyle];


export const MobileSwingCreateStyle = {
  quick: 'quick',
  ride: 'ride',
} as const;

export type MobileSwingCreate = MobileSwingSettings & {
  mode: MobileSwingCreateMode;
  /**
     * @minimum 0.01
     * @maximum 10
     */
  budgetSol: number;
  /**
     * @minimum 12
     * @maximum 720
     */
  windowHours: number;
  universe?: MobileSwingCreateUniverse;
  style?: MobileSwingCreateStyle;
  unlimitedWindow?: boolean;
  targetMint?: string;
  manualBuy?: boolean;
  reentry?: boolean;
  /**
     * @minimum 3
     * @maximum 500
     */
  takeProfitPct?: number;
};

export type AlphaOversightAuthorityEngine = typeof AlphaOversightAuthorityEngine[keyof typeof AlphaOversightAuthorityEngine];


export const AlphaOversightAuthorityEngine = {
  'rules-based-fallback': 'rules-based-fallback',
} as const;

export interface AlphaOversightAuthority {
  ownerWallet: string;
  advisoryOnly: true;
  engine: AlphaOversightAuthorityEngine;
}

export interface AlphaBriefAuthorization {
  /** Unix milliseconds. Sign paif-alpha-oversight|v1|request or history|ownerWallet|nonce. Valid for five minutes. */
  nonce: number;
  /**
     * @minLength 80
     * @maxLength 100
     */
  signature: string;
}

export type AlphaOversightEvidenceArea = typeof AlphaOversightEvidenceArea[keyof typeof AlphaOversightEvidenceArea];


export const AlphaOversightEvidenceArea = {
  market: 'market',
  paper: 'paper',
  payments: 'payments',
  release: 'release',
} as const;

export type AlphaOversightEvidenceStatus = typeof AlphaOversightEvidenceStatus[keyof typeof AlphaOversightEvidenceStatus];


export const AlphaOversightEvidenceStatus = {
  observed: 'observed',
  degraded: 'degraded',
  unknown: 'unknown',
} as const;

export interface AlphaOversightEvidence {
  area: AlphaOversightEvidenceArea;
  source: string;
  status: AlphaOversightEvidenceStatus;
  observedAt: string;
  /**
     * @maxItems 6
     * @items.maxLength 240
     */
  facts: string[];
  /**
     * @maxItems 4
     * @items.maxLength 240
     */
  limitations: string[];
}

export type AlphaOversightBriefEngine = typeof AlphaOversightBriefEngine[keyof typeof AlphaOversightBriefEngine];


export const AlphaOversightBriefEngine = {
  'rules-based-fallback': 'rules-based-fallback',
} as const;

export type AlphaOversightBriefRiskLevel = typeof AlphaOversightBriefRiskLevel[keyof typeof AlphaOversightBriefRiskLevel];


export const AlphaOversightBriefRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface AlphaOversightBrief {
  id: string;
  generatedAt: string;
  advisoryOnly: true;
  engine: AlphaOversightBriefEngine;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  riskLevel: AlphaOversightBriefRiskLevel;
  /** @maxLength 600 */
  recommendation: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  evidence: AlphaOversightEvidence[];
}

export interface HealthStatus {
  status: string;
}

export type UpdateMobileSwingSettings200 = {
  strategy: MobileSwingSummary;
};

export type RespondMobileSwingSignal200 = { [key: string]: unknown };