# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /paif-alpha/briefs:
    post:
      operationId: requestAlphaBrief
      summary: Request a server-generated advisory-only oversight brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlphaBriefAuthorization"
      responses:
        "201":
          description: Durable advisory brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlphaOversightBrief"
        "400":
          description: Invalid input
        "401":
          description: Invalid or expired owner signature
        "409":
          description: Authorization already used
        "429":
          description: Brief request limit reached
        "503":
          description: Durable storage unavailable
  /paif-alpha/briefs/history:
    post:
      operationId: getAlphaBriefHistory
      summary: Read the latest 20 briefs with a separate owner signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlphaBriefAuthorization"
      responses:
        "200":
          description: Newest first, bounded to 20 briefs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlphaOversightBrief"
        "400":
          description: Invalid input
        "401":
          description: Invalid or expired owner signature
        "409":
          description: Authorization already used
        "503":
          description: Durable storage unavailable
...
components:
  schemas:
...
    AlphaBriefAuthorization:
      type: object
      additionalProperties: false
      required: [nonce, signature]
      properties:
        nonce: { type: integer, description: "Unix milliseconds. Sign paif-alpha-oversight|v1|request or history|ownerWallet|nonce. Valid for five minutes." }
        signature: { type: string, minLength: 80, maxLength: 100 }
    AlphaOversightEvidence:
      type: object
      required: [area, source, status, observedAt, facts, limitations]
      properties:
        area: { type: string, enum: [market, paper, payments, release] }
        source: { type: string }
        status: { type: string, enum: [observed, degraded, unknown] }
        observedAt: { type: string, format: date-time }
        facts:
          type: array
          maxItems: 6
          items: { type: string, maxLength: 240 }
        limitations:
          type: array
          maxItems: 4
          items: { type: string, maxLength: 240 }
    AlphaOversightBrief:
      type: object
      required: [id, generatedAt, advisoryOnly, engine, confidence, riskLevel, recommendation, evidence]
      properties:
        id: { type: string }
        generatedAt: { type: string, format: date-time }
        advisoryOnly: { type: boolean, const: true }
        engine: { type: string, enum: [rules-based-fallback] }
        confidence: { type: integer, minimum: 0, maximum: 100 }
        riskLevel: { type: string, enum: [low, medium, high] }
        recommendation: { type: string, maxLength: 600 }
        evidence:
          type: array
          minItems: 4
          maxItems: 4
          items:
            $ref: "#/components/schemas/AlphaOversightEvidence"
...