# lib/api-zod/src/generated/types/alphaOversightBrief.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlphaOversightBriefEngine } from './alphaOversightBriefEngine';
import type { AlphaOversightBriefRiskLevel } from './alphaOversightBriefRiskLevel';
import type { AlphaOversightEvidence } from './alphaOversightEvidence';

export interface AlphaOversightBrief {
  id: string;
  generatedAt: Date;
  advisoryOnly: true;
  engine: AlphaOversightBriefEngine;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  riskLevel: AlphaOversightBriefRiskLevel;
  /** @maxLength 600 */
  recommendation: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  evidence: AlphaOversightEvidence[];
}