# lib/api-zod/src/generated/types/alphaOversightEvidence.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlphaOversightEvidenceArea } from './alphaOversightEvidenceArea';
import type { AlphaOversightEvidenceStatus } from './alphaOversightEvidenceStatus';

export interface AlphaOversightEvidence {
  area: AlphaOversightEvidenceArea;
  source: string;
  status: AlphaOversightEvidenceStatus;
  observedAt: Date;
  /**
     * @maxItems 6
     * @items.maxLength 240
     */
  facts: string[];
  /**
     * @maxItems 4
     * @items.maxLength 240
     */
  limitations: string[];
}